--- Assets/Scripts/GlitchyObject.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 51a3f0d4277863e4eaea4946936c1be8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
