+++ Assets/Scripts/GlitchySprite.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 759f8f87d662bc349aa3ca907f6b218e
+guid: 1fde4e52e809af144948bd074d9ca23e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
