--- Assets/Scripts/NewGlitchySprite.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1fde4e52e809af144948bd074d9ca23e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
