--- Assets/Scripts/Platformer/PlatformerObject.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6e424ea02abb8464e9aeaac5b9749f56
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
