--- Assets/Scripts/Platformer/PlatformerSprite.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f76f4096df5151048b370fa70b6c4260
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
