--- Assets/Scripts/Spawner.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7fa1d0f0a8f6a134490601b9dac27644
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
