--- Assets/Scripts/Tanks/TanksObject.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e3320be9f405b0044a841be188052ddb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
