--- Assets/Scripts/Tanks/TanksSprite.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5eb965bdfe6df174c827c777d2ab2ed9
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
