--- Assets/Scripts/Tinter.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7401f92817ac7fb4f871772e4746fdf3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
