--- Assets/Scripts/space/SpaceObject.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a3649d6f0fcf89e45a29e12c47dd78e6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
