--- Assets/Scripts/space/SpaceSprite.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 01ad4c7b1ebf9ab4a9aaebb871fe30d5
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
